+++ common/compose/src/jsMain/kotlin/dev/inmo/micro_utils/common/compose/OpenLink.kt
@@ -1,0 +1,10 @@
+package dev.inmo.micro_utils.common.compose
+
+import org.jetbrains.compose.web.attributes.ATarget
+
+fun openLink(link: String, mode: ATarget = ATarget.Blank, features: String = "") = dev.inmo.micro_utils.common.openLink(
+    link,
+    mode.targetStr,
+    features
+)
+